requirements.txt
fastapi>=0.110
pydantic>=2.0
uvicorn>=0.27
pyyaml>=6.0
